Time: 5 ms (45.47%) | Memory: 69.8 MB (13.66%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -17,6 +17,6 @@ class Solution {
         maxsum=Math.max(maxsum,sum);
          
         }
-    return maxsum/k;
+    return (double)maxsum/k;
     }
 }
